Split React CDN URL building out of the package registry
`package-registry.ts` held two concerns that never met. Most of it owns the
dependency-snapshot machinery — reading package.json, hashing dependency maps,
remembering snapshots by cache key. Bolted onto that were `esmShReact`,
`getReactUrls`, and `getReactImportMap`, which build esm.sh URLs for React and
touch none of it.

They now live in `src/transforms/esm/react-cdn.ts`, along with the
`DEFAULT_REACT_VERSION` re-export their callers pair them with. The functions
move verbatim; they were already thin adapters over
`import-rewriter/url-builder.ts`, and this commit does not change how a React
URL is spelled or introduce a fourth place that decides.

Fourteen call sites were updated rather than served by a re-export from the
old home. That is the point of the split — a re-export would leave
`package-registry` still named as the place React URLs come from, which is the
thing that made the file read as two files.

`CSSTYPE_VERSION` and `TAILWIND_VERSION` were re-exported here too and turned
out to have no consumers going through this module: the only importer,
`src/cache/config-hash.ts`, already reads them from `url-builder.ts` directly.
The re-export line is gone with them.

`getDependencyPinningSnapshotSync` is now module-private rather than deleted.
The survey behind this change recorded it as having zero non-test callers; it
has four, all inside this file, on the paths that resolve a requested snapshot
against remembered ones. What it did have was a single caller outside the
module, in `rsc/orchestrators/handler.test.ts`, which used it to fetch a
remembered snapshot as a fixture. That now calls
`resolveRequestedDependencyPinningSnapshot`, the public accessor the rest of
the codebase uses for exactly this, and the sync variant is an implementation
detail again.

The six snapshot accessors keep their current semantics; nothing about how a
snapshot is resolved or remembered is touched here.

Tested:
  deno task typecheck
  deno lint src/
  deno task fmt:check
  deno task lint:module-boundaries
  deno task lint:dependency-boundaries
  deno task lint:style
  deno task lint:barrel-jsdoc
  deno task lint:wildcard-exports
  deno test --no-check --allow-all src/server/ src/transforms/ src/html/ src/modules/ src/react/compat/ src/release-assets/ src/build/   # 468 passed (6316 steps), 0 failed

Not-tested: Full repository test suite. `deno task verify:quick` aborts on a
pre-existing `docs:validate` failure (`@example` missing on
src/release-assets/index.ts), reproduced on the base commit, so gates were run
individually.
