+++ src/react/compat/config-generator.ts
@@ -2,7 +2,7 @@ import { rendererLogger as logger } from "#veryfront/utils";
 import { join } from "#veryfront/compat/path/index.ts";
 import { createError, toError } from "#veryfront/errors";
 import { createFileSystem } from "#veryfront/platform/compat/fs.ts";
-import { getReactUrls } from "#veryfront/transforms/esm/package-registry.ts";
+import { getReactUrls } from "#veryfront/transforms/esm/react-cdn.ts";
 import {
   REACT_VERSION_17,
   REACT_VERSION_18_2,
