+++ src/transforms/esm/http-bundler.ts
@@ -8,7 +8,7 @@
 import { rendererLogger as logger } from "#veryfront/utils";
 import type { Plugin } from "veryfront/extensions/bundler";
 import { replaceSpecifiers } from "./lexer.ts";
-import { DEFAULT_REACT_VERSION, getReactUrls } from "./package-registry.ts";
+import { DEFAULT_REACT_VERSION, getReactUrls } from "./react-cdn.ts";
 import {
   type EnvironmentConfig,
   getEnvironmentConfig,
