+++ src/transforms/esm/package-registry.ts
@@ -20,16 +20,7 @@ import type { VeryfrontConfig } from "#veryfront/config";
 import { getHostEnv } from "#veryfront/platform/compat/process.ts";
 import { DEPENDENCY_PINNING_ENV_FLAG } from "../../release-assets/constants.ts";
 import { isExactSemver } from "./npm-registry-client.ts";
-import {
-  buildReactUrl,
-  CSSTYPE_VERSION,
-  DEFAULT_REACT_VERSION,
-  getReactImportMap as getReactImportMapFromRewriter,
-  TAILWIND_VERSION,
-} from "../import-rewriter/url-builder.ts";
-
-// Re-export constants from unified source
-export { CSSTYPE_VERSION, DEFAULT_REACT_VERSION, TAILWIND_VERSION };
+import { DEFAULT_REACT_VERSION } from "../import-rewriter/url-builder.ts";
 
 interface CachedDependencyVersions {
   mtimeMs: number | null;
@@ -270,45 +261,6 @@ export function normalizeReactVersion(version: string | undefined): string {
 }
 
 /**
- * Build esm.sh URL with deps=csstype for React packages.
- */
-export function esmShReact(
-  pkg: string,
-  version: string,
-  path = "",
-  external = false,
-): string {
-  return buildReactUrl(
-    pkg as "react" | "react-dom",
-    version,
-    path || undefined,
-    external,
-  );
-}
-
-/**
- * Get React esm.sh URLs with consistent versioning.
- */
-export function getReactUrls(version?: string): Record<string, string> {
-  const v = version ?? DEFAULT_REACT_VERSION;
-  return {
-    react: buildReactUrl("react", v),
-    "react-dom": buildReactUrl("react-dom", v, undefined, true),
-    "react-dom/client": buildReactUrl("react-dom", v, "/client", true),
-    "react-dom/server": buildReactUrl("react-dom", v, "/server", true),
-    "react/jsx-runtime": buildReactUrl("react", v, "/jsx-runtime", true),
-    "react/jsx-dev-runtime": buildReactUrl("react", v, "/jsx-dev-runtime", true),
-  };
-}
-
-/**
- * Get complete React import map for esm.sh.
- */
-export function getReactImportMap(version?: string): Record<string, string> {
-  return getReactImportMapFromRewriter(version ?? DEFAULT_REACT_VERSION);
-}
-
-/**
  * Strip semver range prefixes (^, ~, >=, >, <=, <, =) from a version string.
  */
 export function stripSemverRange(version: string): string {
@@ -499,7 +451,7 @@ function rememberDependencyPinningSnapshot(
   return snapshot;
 }
 
-export function getDependencyPinningSnapshotSync(
+function getDependencyPinningSnapshotSync(
   source: DependencyPinningSourceInput,
   cacheKey: string,
 ): DependencyPinningSnapshot | undefined {
